% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player-ids.R
\name{get_player_ids}
\alias{get_player_ids}
\title{Computes a tidy correlation}
\usage{
get_player_ids()
}
\value{
A data table with MLBAM, First Name, and Last Name
}
\description{
Computes a tidy correlation
}
